ci: scheduled tests are now twice per week instead of nightly
Nightly seems a bit too frequent, compared to the rate of change on our
production branches; furthermore, when this high-statistics job is
running, my (our?) other workflow runs on other `JeffersonLab`-owned
repos are queued and have to wait for this one.

The new run times are Wednesday and Saturday at 03:30 EST / 04:30 EDT
(=0830Z).